backend/pyproject.toml
[project]
name = "kortix"
version = "1.0"
description = "platform to build, manage and train AI Agents"
authors = [{ name = "marko-kraemer", email = "mail@markokraemer.com" }]
readme = "README.md"
license = { text = "Apache-2.0" }
classifiers = [
  "License :: OSI Approved :: Apache-2.0 License",
  "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.11"
dependencies = [
  "python-dotenv==1.0.1",
  "litellm>=1.77.5",
  "click==8.1.7",
  "questionary==2.0.1",
  "requests==2.32.3",
  "packaging==24.1",
  "setuptools==75.3.0",
  "pytest==8.3.4",
  "pytest-asyncio==0.24.0",
  "pytest-cov==6.0.0",
  "pytest-env==1.1.5",
  "pytest-mock==3.14.0",
  "pytest-xdist==3.3.0",
  "pytest-timeout==2.3.1",
  "pytest-randomly==3.12.0",
  "pytest-rerunfailures==10.2.0",
  "asyncio==3.4.3",
  "altair==4.2.2",
  "prisma==0.15.0",
  "fastapi==0.115.12",
  "uvicorn==0.27.1",
  "python-multipart==0.0.20",
  "redis==5.2.1",
  "upstash-redis==1.3.0",
  "supabase==2.17.0",
  "asyncpg>=0.29.0",  # PostgreSQL async driver for local database
  "pyjwt==2.10.1",
  "exa-py==1.9.1",
  "e2b-code-interpreter==1.2.0",
  "certifi>=2025.4.26",
  "python-ripgrep==0.0.6",
  "daytona-sdk>=0.115.0",
  "daytona-api-client>=0.115.0",
  "daytona-api-client-async>=0.115.0",
  "boto3>=1.40.74",
  "openai>=1.99.5",
  "nest-asyncio==1.6.0",
  "vncdotool==1.2.0",
  "tavily-python==0.5.4",
  "pytesseract==0.3.13",
  "stripe==11.6.0",
  "dramatiq==1.18.0",
  "dramatiq[redis]==1.18.0",
  "phonenumbers==8.13.50",
  "prometheus-client==0.21.1",
  "langfuse==2.60.5",
  "Pillow>=10.4.0",
  "mcp==1.9.4",
  "aiohttp==3.12.0",
  "email-validator==2.0.0",
  "mailtrap==2.0.1",
  "sentry-sdk[fastapi]==2.29.1",
  "gunicorn>=23.0.0",
  "cryptography>=41.0.0",
  "apscheduler>=3.10.0",
  "croniter>=1.4.0",
  "structlog==25.4.0",
  "PyPDF2==3.0.1",
  "python-docx==1.1.0",
  "openpyxl==3.1.2",
  "chardet==5.2.0",
  "PyYAML==6.0.1",
  "composio>=0.8.0",
  "python-pptx>=1.0.0",
  "beautifulsoup4>=4.12.0",
  "weasyprint>=63.0",
  "cssutils>=2.9.0",
  "fastapi-sso>=0.9.0",
  "daytona>=0.21.6",
  "google-api-python-client>=2.120.0",
  "google-auth>=2.28.0",
  "google-auth-oauthlib>=1.2.0",
  "google-auth-httplib2>=0.2.0",
  "typing-extensions==4.15.0",
  "svglib==1.5.1",
  "chunkr-ai>=0.3.7",
  "reportlab>=4.0.0",
  "freestyle>=0.0.17",
  "anthropic>=0.69.0",
  "novu-py>=3.11.0",
  "psutil>=5.9.0",
  "google-analytics-data>=0.18.0",
  "realitydefender>=0.1.10",
  "docker>=7.0.0",
  "asyncpg>=0.29.0",
  "sqlalchemy>=2.0.0",
  "alembic>=1.13.0",
]

[project.urls]
homepage = "https://www.kortix.com"
repository = "https://github.com/kortix-ai/suna"

[tool.uv]
package = false

[dependency-groups]
dev = [
    "orjson>=3.11.1",
]
